+++ bmarks/microbench/linkedlistpgas/linkedlist.h
@@ -51,7 +51,7 @@
 #define TRANSACTIONAL                   d->unit_tx
 
 typedef intptr_t val_t;
-typedef unsigned int nxt_t;
+typedef size_t nxt_t;
 
 #define VAL_MIN                         INT_MIN
 #define VAL_MAX                         INT_MAX
